refactor: クラスの共通化
